Fix unit tests failure due to time racing
The tests passed on the local debugging session but failed during
running. There are 2 issues contribute to the failure on the sever.

1. Initialize() will return slightly earlier than
traceEventSession.Process() is fully running. Which leads to the event
lost for the unit test because the unit test invokes the event
immediately after calling the initilize() on the module.
Although the right fix would be locking on initilize() until Process()
is fully up, it is an overhead to implement. At the same time, in real
world scenario, events are emitted in different process and will be
capatured once Process() is up.
So, I am fixing the issue by simple put a delay in the unit test after
calling initialize().

2. When using F5 (Debug) to invoke unit tests and break the test case in
the middle, the trace event session will be left. That's why after
Debug, the unit tests start to pass because there are existing opening
sessions writing the events into the channel. This introduces huge
inconsistence for the unit test result.
I added clean up logic for the tests to solve this issue.