Document excludeTypes and includedTypes parameters for UseAdaptiveSampling
The `excludeTypes` and `includedTypes` parameters are `string`s with no information what a "type" can be. This PR extends the xmldoc for the parameters with info on the possible values: Dependency, Event, Exception, PageView, Request, Trace

This list is documented on the `SamplingTelemetryProcessor` class. The actual list of values exists in the internal class `SamplingIncludesUtility`.